+++ app/src/main/java/com/sosialite/solite_pos/data/source/local/entity/helper/RecapData.kt
@@ -2,5 +2,5 @@ package com.sosialite.solite_pos.data.source.local.entity.helper
 
 data class RecapData(
     var name: String,
-    var total: Int
+    var total: Long
 )
